38243: Content-tab of category Category and Repository group

+++ Modules/GroupReference/classes/class.ilObjGroupReferenceAccess.php
@@ -30,7 +30,7 @@ class ilObjGroupReferenceAccess extends ilContainerReferenceAccess
             case 'visible':
             case 'read':
                 include_once './Modules/GroupReference/classes/class.ilObjGroupReference.php';
-                $target_ref_id = ilObjGroupReference::_lookupTargetRefId($obj_id);
+                $target_ref_id = (int) ilObjGroupReference::_lookupTargetRefId($obj_id);
 
                 if (!$target_ref_id || !$ilAccess->checkAccessOfUser($user_id, $permission, $cmd, $target_ref_id)) {
                     return false;
